backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend API for Cryage crypto agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno",
    "fastapi",
    "uvicorn",
    "python-binance",
    "pinecone-client",
    "google-genai",
    "sentence-transformers",
    "pydantic",
    "python-multipart",
    "httpx",
    "numpy",
    "pandas",
    "scikit-learn",
]